Add files via upload
A program which asks you to input scores and adds the scores to a list and stops adding to the list when u enter anything which is not an integer. Then outputs the scores entered, their sum, minimum and maximum scores, mean, square of each score, the variance for the sample data given and the standard deviation.